requirements.txt
Flask
gunicorn
google-api-python-client
google-cloud-secret-manager
google-auth
